zodiac-ReadMeUpdate.md
<h2>hello this is test </h2>

# NYT

** link to project **

![preview](week1/classwork/materials/NY Times/images/preview.png)

### How It’s Made
In this replica of the New York Times , I used a bunch of sections to separate all parts of my pages as single elements rather than grouped together. Which made it easier for me to target them and style them in the CSS. 
** Langs used **
HTML & CSS
### optimizations

### Lessons Learned
If I did want to group a bunch of elements I would then give them a class in the HTML & target that class in the CSS instead of writing the same rules fir separate elements. That is my favorite things about CSS. Realizing what things will have simile styling and grouping then together making the code less dense.